fix: MerklePathSet depth consistency
Prior to this commit, the [MerklePathSet] was incorrectly incrementing
the depth by 1 for a given path. The [MerklePath] length will always be
the target depth, ranging from the sibling of the leaf until the node
immediately before the root.

The method [MerklePathSet::update_leaf] was also removed. The
[MerklePathSet] is expected to have the same root for all paths, and
such update was breaking that consistency. It is impossible to update a
Merkle path and maintain the same root. Also, a [MerklePathSet] will not
necessarily contain all the paths that composes the provided root. This
means it is impossible to compute the new root as some of the branches
might be absent.

Finally, the internals of the [MerklePathSet] were simplified. It will
contain a single map from an index to a leaf value+path pair.